Move descendant_get from Object to Class
This is technically a breaking change, but it really wasn't possible to
call this on Objects previously.  Most callers called this on Class.

Fixes #62
