+++ lib/more_core_extensions/core_ext/object.rb
@@ -1,3 +1,2 @@
 require 'active_support/core_ext/object/blank'
 require 'more_core_extensions/core_ext/object/namespace'
-require 'more_core_extensions/core_ext/object/descendants'
